@@ -1,6 +1,6 @@
 {
   "name": "onebacklog",
-  "version": "1.25.0",
+  "version": "1.26.1",
   "private": true,
   "scripts": {
     "dev": "vite",
@@ -9,28 +9,24 @@
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/vue-fontawesome": "^2.0.10",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/vue-fontawesome": "^3.0.2",
+    "@popperjs/core": "^2.11.6",
     "axios": "^1.5.0",
     "bootstrap": "^4.6.2",
     "bootstrap-vue": "^2.23.1",
-    "popper.js": "^1.16.1",
-    "vue": "^2.7.14",
-    "vue-multipane": "^0.9.5",
-    "vue-router": "^3.6.5",
-    "vue2-editor": "^2.10.3",
-    "vuedraggable": "^2.24.3",
-    "vuex": "^3.6.2"
+    "mitt": "^3.0.1",
+    "quill": "^1.3.7",
+    "vue": "^3.3.7",
+    "vue-router": "^4.1.6",
+    "vuedraggable": "^4.1.0",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^1.10.2",
-    "eslint": "^8.49.0",
-    "eslint-plugin-vue": "^8.7.1",
-    "node-sass": "^9.0.0",
-    "sass": "~1.64.2",
-    "vite": "^2.9.16",
-    "vite-plugin-vue2": "2.0.3"
+    "@vitejs/plugin-vue": "^4.4.0",
+    "@vue/compat": "^3.3.8",
+    "vite": "^4.5.0"
   },
   "eslintConfig": {
     "root": true,
