@@ -2,10 +2,9 @@ import { SEV, LEVEL } from '../../../constants.js'
 import { getSprintNameById } from '../../../common_functions.js'
 import { utilities } from '../../mixins/generic.js'
 import commonContext from '../common_context.js'
-import { eventBus } from '../../../main.js'
 
 function created() {
-	eventBus.$on('context-menu', (node) => {
+	this.eventBus.on('context-menu', (node) => {
 		this.showContextMenu(node)
 	})
 }
