@@ -3,14 +3,12 @@
     <h3 v-if="idx === 0">{{ title }}</h3>
     <hr />
     <template v-if="idx >= 0 && draggables.length === 0">
-      <b-button
-        @click="showModal = !showModal"
-        block
-        squared
-        variant="outline-secondary"
-      >Click to create a task here</b-button>
+      <b-button @click="showModal = !showModal" block squared variant="outline-secondary">Click to create a task
+        here</b-button>
     </template>
     <div class="b-card-body">
+      <!-- draggable not fit for compat mde: see https://github.com/SortableJS/vue.draggable.next/pull/152 
+      See also https://github.com/SortableJS/vue.draggable.next/issues/57-->
       <draggable v-model="draggables" :group="idx.toString()">
         <div v-for="item in tasks" :key="item.id">
           <taskitem
@@ -23,14 +21,19 @@
           ></taskitem>
         </div>
       </draggable>
+      <!-- example from https://github.com/SortableJS/vue.draggable.next/blob/master/example/components/nested/nested-test.vue
+        <draggable v-model="list" group="people" tag="div" @start="drag = true" @end="drag = false" item-key="id">
+        <template #item="{ element }">
+          <div>
+            {{ element.name }}
+          </div>
+          <taskitem :taskitem="element" :storyId="storyId" :storyTitle="storyTitle" :productId="productId" :taskState="taskState"
+            :columnName="title" :item="item"></taskitem>
+        </template>
+      </draggable>-->
     </div>
-    <b-modal
-      v-model="showModal"
-      :ok-disabled="taskTitle.length === 0"
-      @ok="procSelected"
-      @cancel="doCancel"
-      title="Enter task title"
-    >
+    <b-modal v-model="showModal" :ok-disabled="taskTitle.length === 0" @ok="procSelected" @cancel="doCancel"
+      title="Enter task title">
       <b-form-input v-model="taskTitle"></b-form-input>
     </b-modal>
   </div>
@@ -44,6 +47,11 @@ import Draggable from 'vuedraggable'
 import TaskItem from './TaskItem.vue'
 
 export default {
+  // compatConfig: {
+  //   MODE: 3, // opt-in to Vue 3 behavior for this component only
+  //   draggable: true // features can also be toggled at component level
+  //   [Erik] did not work
+  // },
   mixins: [authorization, utilities],
   name: 'TaskColumn',
   props: ['productId', 'storyId', 'storyTitle', 'tasks', 'title', 'taskState', 'idx'],
@@ -52,19 +60,28 @@ export default {
     draggable: Draggable
   },
 
-  data () {
+  data() {
     return {
       showModal: false,
-      taskTitle: ''
+      taskTitle: '',
+
+      list: [
+        { name: 'vue', id: 1 },
+        { name: 'script', id: 2 },
+        { name: 'com', id: 3 },
+      ],
+      drag: false
     }
   },
 
   computed: {
     draggables: {
-      get () {
+      get() {
+        console.log('draggables: this.tasks = ' + JSON.stringify(this.tasks, null, 2)
+        )
         return this.tasks
       },
-      set (tasks) {
+      set(tasks) {
         if (this.haveWritePermission(this.productId, LEVEL.TASK)) {
           this.$store.dispatch('updateTasks', {
             tasks,
@@ -77,11 +94,11 @@ export default {
   },
 
   methods: {
-    procSelected () {
+    procSelected() {
       this.$store.dispatch('boardAddTask', { storyId: this.storyId, taskState: this.taskState, taskId: createId(), taskTitle: this.taskTitle })
     },
 
-    doCancel () {
+    doCancel() {
       this.taskTitle = ''
     }
   }
@@ -92,10 +109,13 @@ export default {
 .b-cards-margin {
   margin-left: 5px;
 }
-.b-card-body > * {
+
+.b-card-body>* {
   min-height: 50px;
 }
+
 .b-card:last-child {
   margin-bottom: 5px;
 }
 </style>
+
