@@ -1,5 +1,5 @@
 import { defineConfig } from 'vite'
-import { createVuePlugin as vue } from "vite-plugin-vue2"
+import vue from "@vitejs/plugin-vue"
 
 const path = require("path")
 
@@ -10,6 +10,7 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
+      vue: '@vue/compat'
     },
   }
 })
